+++ socialhome/federate/tests/test_tasks.py
@@ -1,4 +1,4 @@
-from unittest.mock import patch, call
+from unittest.mock import patch
 
 import pytest
 from django.test import override_settings
@@ -249,18 +249,25 @@ class TestSendFollow(TestCase):
         super().setUpTestData()
         cls.user = UserFactory()
         cls.profile = cls.user.profile
-        cls.remote_profile = ProfileFactory()
+        cls.remote_profile = ProfileFactory(
+            rsa_public_key=get_dummy_private_key().publickey().exportKey(),
+        )
 
-    @patch("socialhome.federate.tasks.handle_create_payload", return_value="payload")
-    @patch("socialhome.federate.tasks.send_document")
+    @patch("socialhome.federate.tasks.handle_send")
     @patch("socialhome.federate.tasks.send_profile")
-    def test_send_follow_change(self, mock_profile, mock_send, mock_payload):
+    @patch("socialhome.federate.tasks.base.Follow", return_value="entity")
+    def test_send_follow_change(self, mock_follow, mock_profile, mock_send):
         send_follow_change(self.profile.id, self.remote_profile.id, True)
         mock_send.assert_called_once_with(
-            "https://%s/receive/users/%s" % (self.remote_profile.handle.split("@")[1], self.remote_profile.guid),
-            "payload",
+            "entity",
+            self.profile,
+            [(generate_diaspora_profile_id(
+                self.remote_profile.handle, self.remote_profile.guid
+            ), self.remote_profile.key)],
         )
-        mock_profile.assert_called_once_with(self.profile.id, recipients=[(self.remote_profile.handle, None)])
+        mock_profile.assert_called_once_with(self.profile.id, recipients=[
+            generate_diaspora_profile_id(self.remote_profile.handle, self.remote_profile.guid),
+        ])
 
 
 class TestSendProfile(SocialhomeTestCase):
@@ -272,42 +279,28 @@ class TestSendProfile(SocialhomeTestCase):
         cls.remote_profile = ProfileFactory()
         cls.remote_profile2 = ProfileFactory()
 
-    @patch("socialhome.federate.tasks.handle_create_payload", return_value="payload")
+    @patch("socialhome.federate.tasks.handle_send")
     @patch("socialhome.federate.tasks._get_remote_followers")
-    @patch("socialhome.federate.tasks.send_document")
     @patch("socialhome.federate.tasks.make_federable_profile", return_value="profile")
-    def test_send_local_profile(self, mock_federable, mock_send, mock_get, mock_payload):
-        mock_get.return_value = [(self.remote_profile.handle, None), (self.remote_profile2.handle, None)]
+    def test_send_local_profile(self, mock_federable, mock_get, mock_send):
+        recipients = [
+            generate_diaspora_profile_id(self.remote_profile.handle, self.remote_profile.guid),
+            generate_diaspora_profile_id(self.remote_profile2.handle, self.remote_profile2.guid),
+        ]
+        mock_get.return_value = recipients
         send_profile(self.profile.id)
-        mock_payload_calls = [
-            call("profile", self.profile, to_user=self.remote_profile),
-            call("profile", self.profile, to_user=self.remote_profile2),
-        ]
-        self.assertEqual(mock_payload_calls, mock_payload.call_args_list)
-        mock_send_calls = [
-            call("https://%s/receive/users/%s" % (self.remote_profile.handle.split("@")[1], self.remote_profile.guid),
-                 "payload"),
-            call("https://%s/receive/users/%s" % (self.remote_profile2.handle.split("@")[1], self.remote_profile2.guid),
-                 "payload"),
-        ]
-        self.assertEqual(mock_send_calls, mock_send.call_args_list)
+        mock_send.assert_called_once_with(
+            "profile", self.profile, recipients,
+        )
 
     @patch("socialhome.federate.tasks.make_federable_profile")
     def test_skip_remote_profile(self, mock_make):
         send_profile(self.remote_profile.id)
         self.assertFalse(mock_make.called)
 
-    @patch("socialhome.federate.tasks.handle_create_payload", return_value="payload")
-    @patch("socialhome.federate.tasks.send_document")
+    @patch("socialhome.federate.tasks.handle_send")
     @patch("socialhome.federate.tasks.make_federable_profile", return_value="profile")
-    def test_send_to_given_recipients_only(self, mock_federable, mock_send, mock_payload):
-        send_profile(self.profile.id, recipients=[(self.remote_profile.handle, None)])
-        mock_payload_calls = [
-            call("profile", self.profile, to_user=self.remote_profile),
-        ]
-        self.assertEqual(mock_payload_calls, mock_payload.call_args_list)
-        mock_send_calls = [
-            call("https://%s/receive/users/%s" % (self.remote_profile.handle.split("@")[1], self.remote_profile.guid),
-                 "payload"),
-        ]
-        self.assertEqual(mock_send_calls, mock_send.call_args_list)
+    def test_send_to_given_recipients_only(self, mock_federable, mock_send):
+        recipients = [generate_diaspora_profile_id(self.remote_profile.handle, self.remote_profile.guid)]
+        send_profile(self.profile.id, recipients=recipients)
+        mock_send.assert_called_once_with("profile", self.profile, recipients)